Changes in the folder structure.
